🚚 그룹 라우팅으로 서로 다른 레이아웃 적용
