+++ src/app/group/(layout)/_components/GroupCard/Description.tsx
