+++ src/app/group/(layout)/_components/GroupCard/Owner.tsx
