+++ src/app/group/(layout)/_components/GroupCard/index.ts
