+++ src/app/group/(layout)/_components/GroupSearch.tsx
